Initial support for deterministic sampling
Added support for deterministic sampling
Sampling should be deterministic for both pre-trianing, as well as with
meta-training. This is done by applying an extra wrapper around task-dataset
which saves the random state at each get-time. Each task-loader is
deterministic for a single epoch and should then be resampled.

TODO: Add tests for this